requirements.txt
Flask==3.0.3
requests==2.32.3
google-api-python-client==2.146.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
mcp==1.13.1   # latest stable version available
